news_trends/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nltk==3.4.5
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
